backend/requirements.txt
Flask
Flask-Cors
PyPDF2
Pillow
reportlab
